Remove the NoErrorsPlugin
It turns out that this plugin causes more trouble than it’s worth.  It
can result in the browser silently failing to update when there’s an
error, and you might not notice if you’re not watching your console.

Failing loudly is preferred to this, even if it requires a page refresh.