Clean up remaining lint violations
In order to fix all of the `import/no-unresolved` lint rule violations,
we need to use `eslint-import-resolver-webpack`.  That plugin allows us
to specify a webpack configuration, but I found some issues with that:

* If I use `webpack/development.js`, then the production build would
fail with CSS errors.
* If I use `webpack/production.js`, then the dev build would fail with
CSS errors.

To solve the problem, I created a `webpack.config.js` (default
filename) that conditionally pulls in the correct configuration
(production or development).

As a bonus, that file allows us to remove most of the explicit
configuration files we had in package.son.