Update eslint-plugin-react to 4.1.0
This brings in some new rules and bug fixes.

Most notable likely effect is that the prop-types rule should now
recognize stateless function components that destructure props in the
argument list, which will force use to define propTypes for a bunch of
components that we haven’t been forced to up til now.