Improve testing infrastructure
Goals:
* Be able to run tests from the command line or from the browser using
code compiled the same way (as much as possible given browser vs node
differences)
* Have an `npm run test:watch` that doesn't thrash the CPU
* Avoid having to work around the absence of webpack loader issues when
running Mocha.

This solution does the following:
* Use the new mocha-webpack package for command-line test running (and
watching).  In my limited testing this seems to solve the mocha --watch
issues with CPU load
* Can run a second webpack-dev-server on 8081 that runs the tests in
the browser with hot reloading.  The hot reloading makes it so that
only tests that were affected by a code change are run.  Webpack's very
nature makes it possible to precisely identify which tests need to be
re-run, resulting in a pretty sweet workflow
* Eliminates the cssModules and shimImageLoader work-arounds
* Eliminates Karma
* Makes some configuration changes to work around various browser
testing issues