Fix spurious lint warning when running tests
Using webpack-node-externals was causing the spurious lint warning
about an unresolved import and doesn’t seem to be needed for any other
purpose, so I removed it.

We also no longer need to specify `NODE_PATH` when running tests,
because webpack does that for us; it was a holdover from when we were
running Mocha as a standalone thing.