Remove scss loader duplication from webpack configs
`development.js`, `test.js`, and `testDebug.js` all shared the same
configuration for loading SCSS files.  `production.js` is similar, but
not quite the same, as it uses the ExtractTextPlugin as well.

So, in `production.js`, we find the base loader entry for scss files
and replace it with the right one for production.  We could just wrap
the base loader but #38 changes the `localIdentName` setting for
development/test, and we don’t want that change in production.