Use css-loader/locals for Mocha tests
We had been using a specific commit of style-loader that matched an
outstanding PR that made it work in the node environment
(https://github.com/webpack/style-loader/pull/115).

That PR was rejected with a suggestion to use `css-loader/locals`
instead of the style-loader, so that’s what I’ve done here, which
allows us to go back to the stock version of style-loader.

I reverted the change that pulled the SCSS handling up to the base
webpack config as there is now much less duplication.

I ran into a bunch of strange behavior along the way, but what’s here
seems to work well:

* The decorated style names have an extra `_-_-` prefix on them that
doesn’t show up in the browser environment.
* Using the trailing `!sass` on the loader definition resulted in a
random suffix of either `—sass` or `-sass` or none at all being added
to the style names, and it’s inconsistent enough to potentially make
specs fail.  Moving the other loaders to a separate element of the
array resolves this.
* I found that I don’t actually seem to need the sass loader at all,
and that makes no sense to me at all.  I’ve left it in because it seems
like it really needs to be there, but I am very confused by what I saw.