Adapt redux-api-middleware to work with redux-form
With `formApiAdapter`, you can wrap a redux-api-middleware action
creator in such a way that it will work as expected by redux-form.  The
promise returned by the adapter will:

* Resolve with a successful payload
* Reject with a failing payload

If the failing payload is a 422 error, the rejection contains error
information in the format expected by redux-form.

The error formatting is Rails-specific for now.  It would be possible
to pass the error formatter into `formApiAdapter` and rename our
current formatter as `railsErrorFormatter` or something like that.