Eliminate the need to import specHelper everywhere
While the [issue I opened on
mocha-webpack](https://github.com/zinserjan/mocha-webpack/issues/1) is
still outstanding, the author’s suggestion is to use
`webpack-node-externals`.  We tried that initially, but that results in
numerous firings of the `import/no-unresolved` lint rule.

After conferring with @lexun, we decided to go back to that approach
and instead disable the linters when doing `npm run test`.  They still
run with `npm run build`, `npm run test:debug`, and `npm run dev`.

If the above-mentioned issue gets resolved, then we can go back to
running the linters in all environments.