Always run eslint in the production environment
When running on CI, `npm run lint` will be run with `NODE_ENV=test`.
This causes the test webpack config to be active, and the
`import/no-unresolved` rules don’t work properly with that config.
This ultimately causes CI builds to fail.

So instead, we now explicitly run the `lint` script with
`NODE_ENV=production`.