Use latest mocha-webpack
The latest mocha-webpack adds a `—include` option that works like
`—require`, but instead includes the file into the webpack bundle.  We
use it for our spec helper file which makes it available to our specs
even with using `—watch`.

Bottom line: We can now include our specHelper in one place; we no
longer need webpack-node-externals, and we can keep the linters turned
on for test runs (we had them disabled before due a conflict with
webpack-node-externals).