Remove git hooks support
The `ghooks` package doesn't work well with GUI git tools.  They're
aware of that and working on the issue (see
https://github.com/gtramontina/ghooks/issues/40), but for now I'm
removing them from this PR.

I'll watch that issue and re-consider once they've resolved it.

Even if you opt out of the hooks, they still have to run just enough to
know to opt out.