Get chai-as-promised working
chai-as-promised registers itself with chai when imported via the Karma
config, but when we do `import { expect } from ‘chai’` in our specs, we
get a different instance of chai (or something like that).  With Karma,
`expect` is available as a global, so we can just remove the `import`s
and everything works.

Because `expect` is now a global, we have to tell ESLint about it.
I’ve done that in each of the test `.eslintrc.json` files for now.  A
better approach would be to add that to eslint-config-zeal/chai.js
instead.  That should be done before finishing off this branch.