Update peer dependences of eslint-config-zeal
It turns out that eslint-plugin-react’s no `no-children-prop` has a bug
when dealing with spread props, and that gets triggered in
`layout/Base/index.js`.

I’ve submitted a PR to that project
(https://github.com/yannickcr/eslint-plugin-react/pull/862), but for
now am pointing at my PR branch.

It would be possible to extract the props to a variable before passing
them to `React.createElement` to work around the bug.  I’m happy to go
that way instead of having a github dependency in the project.

Or, we could let this PR sit open for a few days to see if the PR gets
merged upstream.